AV: add TARGET_IGNORE_VENDOR_AUDIO_EFFECTS_CONF
Devices with a vendor partition that include audio_effects.conf are
hitting this and always using their audio_effects.conf. The issue presented
with this is that we cannot make changes to the file in the vendor partition,
and thus symlinked to system/vendor.

Adding this opt-in will always load the audio_effects.conf file in
/system/etc as well as always ignore the one in vendor/etc (and therefore system/vendor/etc
if the device opted-in to the vendor partition symlink)

Change-Id: Iaaf5f561629c720e303617837e59bb83a395c645
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>